requirements.txt
bip-utils==2.9.3
Flask==3.1.0
pycryptodome=3.21.0